requirements.txt
python-dotenv
spotipy
google
flask
pymongo